Move Factory testing to appropriate directory
