+++ src/test/java/game_entities/TestFactoryClassBoard.java
@@ -1,4 +1,4 @@
-package FactoryClassTesting;
+package game_entities;
 import static org.junit.Assert.*;
 
 import game_entities.Board;
