+++ src/test/java/game_entities/tiles/TestFactoryClassProperty.java
@@ -1,4 +1,4 @@
-package FactoryClassTesting;
+package game_entities.tiles;
 
 import static org.junit.Assert.*;
